requirements.txt

streamlit>=1.20.0
pandas>=1.3.0
python-dotenv>=0.21.0
